requirements.txt
blinker==1.4
click==6.7
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.1
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
SQLAlchemy==1.3.20
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.3.3
